refactor(forecast): update weather Icon component using font awesome package instead of svg
