+++ src/lib/theme.js
@@ -7,8 +7,6 @@ const theme = {
     active: '#51557A',
   },
   filters: {
-    primary:
-      'invert(100%) sepia(1%) saturate(0%) hue-rotate(160deg) brightness(105%) contrast(103%)',
     alternative:
       'invert(75%) sepia(55%) saturate(2317%) hue-rotate(2deg) brightness(107%) contrast(103%)',
   },
